# release-22.1: opt: cache not-null columns in the metadata