Backport 2/2 commits from #86606.

/cc @cockroachdb/release

---

**opt: add a new query to the slow queries benchmark**

Release justification: non-production code changes

Release note: None

**opt: cache not-null columns in the metadata**

This commit adds a new table annotation for not null columns to avoid
rediscovering them multiple times per table. This improves performance
for queries over tables with a large number of columns and indexes.

Benchmark results:
```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-3-10     123ms ± 1%      63ms ± 2%  -48.62%  (p=0.000 n=10+9)
SlowQueries/slow-query-1-10    20.3ms ± 1%    20.2ms ± 2%     ~     (p=0.353 n=10+10)
SlowQueries/slow-query-2-10     312ms ± 2%     313ms ± 4%     ~     (p=0.842 n=10+9)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-3-10    59.3MB ± 0%    49.0MB ± 0%  -17.38%  (p=0.000 n=10+10)
SlowQueries/slow-query-1-10    12.3MB ± 0%    12.3MB ± 0%   -0.27%  (p=0.000 n=10+10)
SlowQueries/slow-query-2-10     181MB ± 0%     181MB ± 0%     ~     (p=0.739 n=10+10)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-3-10      542k ± 0%      381k ± 0%  -29.72%  (p=0.000 n=9+10)
SlowQueries/slow-query-1-10      131k ± 0%      131k ± 0%   -0.39%  (p=0.000 n=10+10)
SlowQueries/slow-query-2-10     1.65M ± 0%     1.65M ± 0%     ~     (p=0.345 n=9+10)
```
Release justification: low risk, high benefit changes to existing
functionality

Release note (performance improvement): Planning time has been reduced
for queries over tables with a large number of columns and/or indexes.